refactor(operator_commands_terminal/read_view): split tests to sibling tests_read_view.rs (#1266)
Inline test module in read_view.rs is moved into sibling tests_read_view.rs
following the existing operator_commands_terminal convention (already has
tests_construction.rs, tests_validation.rs).

- src/operator_commands_terminal/read_view.rs: 464 → 223 LOC (under 400-line rule)
- src/operator_commands_terminal/tests_read_view.rs: new (~241 LOC)
- src/operator_commands_terminal/mod.rs: register mod tests_read_view
- super::* → super::read_view::* (sibling level)
- Added use std::path::PathBuf, use crate::CopilotSubmitAudit (matching tests_validation.rs)

Pure refactor: zero logic changes, all 66 tests pass.

Closes 8/70 of #1266 (this loop iteration), 17 PRs cumulative since #1285.
